requirements.txt
fastapi>=0.104.1
pydantic>=2.5.1
pydantic-settings>=2.1.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
python-dotenv>=1.0.0
geopy>=2.4.1
aiosqlite>=0.17.0  # Для асинхронной работы с SQLite
nbformat>=5.9.2  # Для создания Jupyter notebooks
jupyter>=1.0.0  # Для работы с notebooks
requests>=2.31.0  # Для HTTP запросов
